requirements.txt
absl_py==1.2.0
numpy==1.21.5
ogb==1.3.5
scikit_learn==1.1.3
scikit_learn==1.1.2
torch==1.12.0
torch_geometric==2.1.0.post1
torch_sparse==0.6.15
tqdm==4.64.1
